feat: expose a Bazel rule
Previously we documented that users should use the container_test rule from rules_docker:
https://github.com/bazelbuild/rules_docker/blob/8c3a8110a0c519929a7e79c39ac345a0f8c74d04/contrib/test.bzl#L125
However, rules_docker is no longer maintained. rules_oci means to be a replacement, but doesn't want to add a rule for a binary that is published elsewhere.
It's easier to understand if this repo publishes both the binary, and the bazel rule that wraps it, so that any changes can be made to them together.