Bazel rule: detect "oci_layout" and "oci_tarball" output groups
With rules_oci, the output of "oci_image" and "oci_image_index" is always a tree artifact containing the full oci layout.
rules_img is an upcoming Bazel ruleset for building container images that tries to avoid unnecessary work.
To that end, rules_img will only materialize a full oci layout tree artifact if a downstream action needs it.
This is done through the use of two special output groups:

- oci_layout contains a tree artifact of the full oci layout
- oci_tarball contains is a tar file containing the oci layout

With this change, container_structure_test is compatible with both rulesets.
Users can continue to pass a single file or can choose to use the aforementioned output groups.

This results in a better user experience for rules_img users who otherwise need to create a special intermediate filegroup to convert from OutputGroupInfo to DefaultInfo.